auth

+++ backend/internal/api/responses.go
@@ -1,6 +1,6 @@
 package api
 
 type TokenResponse struct {
-	UserID      int    `json:"userID"`
+	UserID      int64  `json:"userID"`
 	AccessToken string `json:"accessToken"`
 }
